save file as a util
